Refactor: Move favorite handling to ArticleCard
