+++ components/ArticleList/ArticleList.js
@@ -1,14 +1,12 @@
 import React from "react";
 import ArticleCard from "../ArticleCard/ArticleCard";
 
-export default function ArticleList({articles}) {
-  
-
+export default function ArticleList({ articles }) {
   return (
     <ul>
       {articles.map((article) => (
         <li key={article.shareURL}>
-        <ArticleCard article={article} />
+          <ArticleCard article={article} />
         </li>
       ))}
     </ul>
